backend/requirements.txt
# FastAPI backend
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-dotenv==1.0.0

# LangChain stack (compatible with OpenAI 1.40+)
langchain==0.3.7
langchain-core==0.3.18
langchain-community==0.3.7
langchain-openai==0.2.2
langchain-chroma==0.1.4
langchain-text-splitters==0.3.0

# OpenAI API
openai==1.40.6

# Vector DB
chromadb

# Document loaders
pypdf==4.3.1
pandas==2.2.3
python-docx==1.1.2
unstructured==0.15.12
beautifulsoup4==4.12.3
lxml==5.3.0

# Tokenization / embeddings
tiktoken==0.7.0

# Optional for async requests or proxies
httpx